description
Write a query to display the customer name and customer code of all customers who uses 'yahoo' as their email. The customer code should be generated by concatenating the first 3 characters of the customer's name and the first 3 numbers of their contact number. Give an alias name as 'CUSTOMER_CODE'.