description
Write a query to display the customer's name, customer address, and contact information of customers whose address has the 5th letter 'o'. If the customer email is not available,then display the contact number. Give an alias name to the contact information as 'CONTACT'. 

Sort the results based on the customer's name in descending order.