description
Due to some reason the expected delivery date is delayed for a month.

Write a query that display the courier ID, from address, to address and delayed delivery date of all the couriers that booked in the month of 'August'. Give an alias name to the delayed delivery date as 'DELAYED_DELIVERY_DATE'. 

Sort the results based on the to_address in descending order.