description
Write a query to display the courier details such as courier id with the alias name 'COURIER1' and the courier id of the courier that got delivered on the same date as courier1 with the alias name 'COURIER2' and the delivered date. 

Sort the records by the delivery date in descending order and then by the courier id of courier1 in descending order.